+++ assets/shell_completions/borgreport.bash
@@ -19,7 +19,7 @@ _borgreport() {
 
     case "${cmd}" in
         borgreport)
-            opts="-h -V --env-dir --env-inherit --text-to --html-to --metrics-to --mail-to --mail-from --no-progress --glob-archives --check --borg-binary --max-age-hours --help-man --help --version"
+            opts="-h -V --env-dir --env-inherit --text-to --html-to --metrics-to --mail-to --mail-from --no-progress --glob-archives --check --check-options --borg-binary --max-age-hours --help-man --help --version"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 1 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0
@@ -101,6 +101,13 @@ _borgreport() {
                     ;;
                 --check)
                     COMPREPLY=($(compgen -W "true false" -- "${cur}"))
+                    if [[ "${BASH_VERSINFO[0]}" -ge 4 ]]; then
+                        compopt -o nospace
+                    fi
+                    return 0
+                    ;;
+                --check-options)
+                    COMPREPLY=("${cur}")
                     if [[ "${BASH_VERSINFO[0]}" -ge 4 ]]; then
                         compopt -o nospace
                     fi